06_End_to_End_Q&A/requirements.txt
langchain-openai
langchain 
python-dotenv
langchain-community
langchain_core
streamlit
openai 
